requirements.txt
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
factory-boy==3.2.0
